rename sandbox tools.py -> utils.py per review
Renames the module (and its test) and repoints the package re-exports;
public API (execute_code, MCP_SANDBOX_FEATURE_FLAG) is unchanged.
